Attempt bugfix for IOS devices.
Reportedly,
on iOS devices, regardless of browser, the html5 audio player would
reach a hung state when the play button would be clicked.
This was flaky.

Empirically, it was observed that on instances when the hung state
happens, the formatDisplayTime() fn errored out. This was likely due to
the face that the Date::toISOString() was receiving Infinity as an
argument and hence was erroring out.

This fix should account for that case, and hopefully fix the general
hung state issue on IOS.