[BUGFIX]: fix issues with progress bar seek
Key Learnings:
1. crash was happening because of a known case encountered before:
remember to use integers for DateTime.add()

2. As for the dom issues, using the browser's debugger showed that when
using a watch expression for the style state, it would go out of
scope (implying a re-render). Hence the realisation that it's a matter
of ignoring the dom-patching by LiveView. Based on the documentation for
it [1], it seems that ignoring dom-patching should be considered
whenever we use JS-hooks to programmatically manipulate the state.

[1] https://hexdocs.pm/phoenix_live_view/dom-patching.html