Refactor event-patterns, use MediaSessions API (#69)
## Overview

This PR does a bunch of things: 
1. refactor of events bridge 
2. use media sessions api


## Migration Notes:
- most recent db state dump (for migration):
[1723306013568018000.json](https://github.com/user-attachments/files/16570171/1723306013568018000.json)


## Commit Descriptions:


* Cleanups: rm old .livemds

* Update migrator script's save() to preload video

* fix imprt stmt & add guard to hoverrune

* Wire up voice metadata

note: because of this, the PR requires the deployment to use a fresh json file

this prepares for the info needed by media sessions

does some linting using the latest elixir-ls

* Use og image for artwork

Not sure why it doesn't display when using the emulator.
I think it's better if I test it on the deployed version directly, not
sure if the testing method is accurate.

* Minor cleanups

* Meta cleanup encode

* Add comma

* Simple changes

* Revert use of Map.from_struct

* Shift guarding to playback.ex instead of encoder

* Add nil clause for meta::Jason encode()

* @derive

* except artwork

* Handle nil case for artwork better w better guard

* Minor changes

* Use dummy metadata example

* Revert dummy example

* iAnother dummy attempt

* Another attempt

* test dict creation

* test use of extra key in media metadata

* Test: use geenrated artwork

* minor change

* Change MIME type for generated_artwork jpeg -> png

* Minor change

* Use createMediaMetadataPayload()

* Server Logging to Console & Image Lib Update

* Init media session @ metadata load

Init functions don't have action handlers yet

* Guard against null playback

* Shift mediaSession update to before play

* Attempt earlier update of metadata

* Use waiting spinner to prevent early play triggers

* Trigger diffs to create draft PR

* Clean up fns on media_bridge.ex

mainly consolidates the functions used for updating playback and
notifying the audio player

* Shift JS event dispatch to AudioPlayer livecomp

Separates the concerns b/w the MediaBridge (generic) and the
AudioPlayer (concrete)

* Shift generic hook event dispatch to media_bridge

* Minor cleanups on MediaBridge hook

* Refactor event handlers within MediaBridgeHook

* Cleanup some handshakes, msg passing

* Fix: disregard duplicate voice_acks

* Add playbackMetaBridge, loadAudio @ earliest time

Change list:

1. reorganise bridges into hooks/mediaEventBridges, within it there's a
new playbackMetaBridge. Intent is to comms playback info and separate
actions from playback info things.

2. now, when the server does a push_event for the audio events
registration, we also send a separate event that is using this new
bridge ("media_bridge:registerPlayback"), which dispatches the necessary
client side events to register the playback, load the audio and the
mediaSession as early as possible. Previously, this used to happen JIT,
when the user was to press the play/pause button. This seems to have
removed the initial time-delay we used to have between the instance when
user clicks "play" and actual playback starting.

3. naturally, we also prevent redundant re-loads of audio now by
guarding the source-setting for the html5 audio player.

Broad TODOs:
1. consider pushing to the new playbackMetaBridge in the same intervals
as the existing heartbeat.

2. this commit only added in message-passing from MediaBridgeHook to the
AudioPlayerHook, but not in the other direction.

* Rm props-passing of playback via <audio/>

Playback state is init @ point of events registration, via a new bridge now.

* Wire up PlayPause Action handlers for MediaSession

Automagically works == pat on the back for good design?

* Rm ack_val passing, implicitly get from the struct

* Cleanup: rm useless event handlers for audioplayer

* Cleanup: rm console.logs from hooks

* Cleanups: minor cleanups on server components

* minor changes

---------

Co-authored-by: ks0m1c_dharma <sakiyamuni@sams.ara>