Rm use of BrowserNavInterceptor
Initially this was supposed to be an attempt at a workaround, but now
that we've bumped up the version number for live_view, the navigation
state handling based on the back button works okay.

To be honest, the state management is still a little flakey but I have
spent too long investigating this, will move on first.