Add Slottable ActionBar, cohere w MediaBridge
Similar to the hoverune implementation, this action_bar.ex live_component also takes in an array of actions, which the user_mode defines and then renders buttons that trigger click events specific to the mode and action_name.

This allows the DM to be the one with the necessary handlers to act on these events.

******* 1. Bars are soft appearing / disappearing
This means that we don't mount / destroy nodes, simply hide and make them appear as and when we need.
This means that we need to make some things "equivalent" and work the same way as it does on MediaBridge and ActionBar.

Suppose there are nav buttons (the left and right stuff) which are generic.
They will DEFINITELY appear on the MediaBridge.
They may or may not appear on the ActionBar.

This means that any button click on the nav buttons on MediaBridge may just be relayed to a optionally present button on the actionbar, along with other buttons on the action bar.

******* 2. Maintain equivalence in UI buttons b/w action-bar and media bridge by relaying button clicks
This will Shim / Relay the nav button click from the media bridge to the corresponding action bar via hooks

Essentially, clicking the nav buttons on the mediabridge will relay to the nav buttons on the action bar, which is mode-specific
Here's some logs to show that the wire-up looks okay (have left the functions empty first).

%Vix.Vips.Image{ref: #Reference<0.3949460734.1140719640.214139>}
[debug] HANDLE EVENT "read::nav_back" in VyasaWeb.DisplayManager.DisplayLive
  Parameters: %{"value" => ""}
TRACE: TODO handle nav_event @ action-bar region: %{"event" => "nav_back", "mode" => "read"}
[debug] Replied in 430µs
TRACE: TODO handle nav_event @ action-bar region: %{"event" => "nav_fwd", "mode" => "read"}
[debug] HANDLE EVENT "read::nav_fwd" in VyasaWeb.DisplayManager.DisplayLive
  Parameters: %{"value" => ""}
[debug] Replied in 375µs
TRACE handle event for toggle_show_control_panel: false
[debug] HANDLE EVENT "toggle_show_control_panel" in VyasaWeb.DisplayManager.DisplayLive
  Component: VyasaWeb.ControlPanel
  Parameters: %{"value" => ""}
[debug] Replied in 342µs
[debug] HANDLE EVENT "change_mode" in VyasaWeb.DisplayManager.DisplayLive
  Parameters: %{"current_mode" => "read", "target_mode" => "draft", "value" => ""}
[debug] Replied in 162µs
[debug] HANDLE EVENT "draft::nav_fwd" in VyasaWeb.DisplayManager.DisplayLive
  Parameters: %{"value" => ""}
TRACE: TODO handle nav_event @ action-bar region: %{"event" => "nav_fwd", "mode" => "draft"}
[debug] Replied in 562µs
TRACE: TODO handle nav_event @ action-bar region: %{"event" => "nav_back", "mode" => "draft"}
[debug] HANDLE EVENT "draft::nav_back" in VyasaWeb.DisplayManager.DisplayLive
  Parameters: %{"value" => ""}
[debug] Replied in 573µs
iex(test@127.0.0.1)1>