Init VyasaWeb.Content.ReadingContent
This is a big commit, commit messages to signpost will be added to this.

Essentially, we want DM to be a mediator and therefore all the
mode-specific logic should be taken out of the DM and into the
mode-specific content module.

ReadingContent is the component that shall be slotted in for the "read"
mode.

This particular commit is not a complete refactor, but takes the
following steps towards that direction:
1. shifts the read-mode-specific data loading to the content module.
this means all the apply actions have been updated.

2. adhere to the pattern where content livecomponent is it's own source
of truth (this part is not fully refactored yet, so DM still injects
some state into the content livecomponent). This particular pattern
seems to actually follow [known Phoenix patterns actually](https://hexdocs.pm/phoenix_live_view/Phoenix.LiveComponent.html#module-livecomponent-as-the-source-of-truth).
    * the liveview only prop-passes the id
    * comms b/w liveview and livecomponent happen using =send_update()=, this is why the =id= for the livecomponent is imporant

3. update event emitters within HEEX strings such that all the events
are targetted. For example, the "clickVerseToSeek" will target dom id =
`#reading-content`. This adheres to the pattern where mode-specific
functionalities happen within their respective components. In this
example, VerseMatrix is a nested child of verse, which is a child of the
reading_content live component. Since this particular event has nothign
to do with other generic slots (e.g. actionBar), then there's no need
for DM to be involved with the handling of this event.

4. DM shall keep things generic, Content is a new slot, and arguments
for it shall be provided by the

5. NOTE: the MediaBridge will be reworked in the next few PRs, which
would simplify how the handshakes happen. For now, this commit actually
introduces some regression bug in the handshake actually. We could spend
a short 30min trying to debug it, but it should be alright to ignore
this regression bug until we get to the point where MediaBridge is
getting refactored, during which we can fix it altogether.

Next Steps:
1. complete point 1 above, there's still the hoverrune and other
read-mode specific events that are yet to be ported.