Add UI improvements to EditableMarkDisplay
The edit button on quick draft should only be seen when the dropdown is
expanded and all the marks can be seen. The general pattern I'm trying
to follow is that buttons should only be shown when an action makes
sense. So here, if the user can't even see the marks currently
accumulating, then the user shouldn't have the chance to edit any marks.

Notice that the EditableMarkDisplay is a live_component that is a child
of a function component (collapsible_marks_display). This is so that it can keep its own state and
do it's own event-handling. I'm not sure if this is pattern-breaking.
A reference to the actual parent live_component that calls collapsible_marks_display
will always be there in case there's a need for the parent to handle
any particular events. So comms b/w live_components is still going to be
via phx-target values.