Use amortized deletes on marks
This commit contains a db migration, so pls do `mix ecto.migrate`

Things done:
1. the enum wasn't created in the db before (checked via pgadmin as
well), it's now created and is part of the latest migration file that
comes with this commit
2. The deletion from the db is done correctly, the helper function for
mutation does the sanitation is doing the contiguous order values
correctly, but it's NOT being persisted, I don't have a theory of mind
on why.
[BUG]: the updated order values are NOT being persisted.
3. [BUG]: regardless the bug in 2., the livecomponent should update,
it's not happening right now, not sure why also

QQ: Also it's not exactly amortized deletes on marks, since each time we do
a handle on create / delete of marks, the mutation of the reflector
happens and the db is written to, so it's actually JIT, no?