Standardise use of @event_target, cleanups
When using @event_target, the intent is that phx-target values
will be passed around. This must be agnostic to the 2 main ways of
supplying that argument, as a HTML selector, or as a module name (or @myself).
Therefore, we should discourage ourselves from doing format-strings at
the point of the @event_target being supplied as an argument, meaning:

good, because agnostic:
phx-target={@event_target}

bad, because hardcoded:
phx-target={"#" <> @event_target}
* in this case, the event trigger will throw a dom error.