[BUGFIX] Explicitly pipe apply_action, not helm
This fixes the error where our previous way of doing a re-direct wasn't
working.
The requirement is that for sources without multiple chapters, the sole
chapter should be loaded directly.

There are 2 ways to go about doing this:

A) do a redirect (push-patch) on this scenario

B) since it's all handled by the apply_actions within the read context,
pipe the execution into the apply action of :show_verses.

This commit favours B over A. It's unclear why A doesn't work.