Better, granular %SheafUiState management
here's my implementation notes

We now have a nice commit_marks_in_reflector. We just want to effect the
DB writes here so that shall be its main responsiblity.

Potentially, this also happens to be a good place to update the ui,
since the draft reflector will be set here. However, doing so actually
fattens the function up because it does ui state changes as well.

To determine if the roles of this function should be changed, we
consider the current callign contexts for the function:
1. when doing an edit mark content ([ref])
   ⁃ DONE ui should do a ui_toggle_is_editing_mark_content()
2. when creating a mark ([ref])
   ⁃ ui should insert this new entry to the ui map that’s in the marks
     ui state ui_add_mark()
3. when creating a mark on the same binding ([ref])
   ⁃ same as 2.
4. when tombing a mark ([ref])
   ⁃ ui_remove_mark()

The ui change should be happening outside of this commit function. It
will be called from the handle_events, we can expose wrapper functions
within the read context and within the ui state to handle these.

Next, The commit function also calls a maybe_prepend_draft_reflector and
this needs to be updated too. ([ref]) it should also do an addition just
like before.

The best way to do this is to just expose register/deregister functions
at various levels of abstraction.

[ref]
<file:~/Projects/vyasa/lib/vyasa_web/components/contexts/read.ex::def
handle_event(>

[ref]
<file:~/Projects/vyasa/lib/vyasa_web/components/contexts/read.ex::"mark::createMark",>

[ref]
<file:~/Projects/vyasa/lib/vyasa_web/components/contexts/read.ex::"mark::tombMark",>

[ref]
<file:~/Projects/vyasa/lib/vyasa_web/components/contexts/read.ex::def
maybe_prepend_draft_mark_in_reflector(>