marks: ui:toggle, and editing in published sheafs
For the marks edit wiring here, it's in the simpler case of marks that
are not in the drafting context (not in the draft sheaf).

This refers to existing marks that are NOT in the drafting context. It's a simpler editing case that is a subset of a generic edit that handles both draft and published sheaf editing. The reason it's simpler is because there's no re-ordering or insertion actions being done, so technically no need to sanitise the marks before updating the sheaf.